Update From Last Build

+++ Year2GroupProject/Assets/User Interface/Scripts/HUD/HUDHealth.cs
@@ -20,7 +20,7 @@ public class HUDHealth : MonoBehaviour {
 
 	void OnGUI () {
 		currentHealth = playerHealth.getPlayerHealth ();
-		percentageHealth = currentHealth / initialHealth; 
+		percentageHealth = (float) currentHealth / initialHealth; 
 		healthBar.fillAmount = percentageHealth;	
 	}
